Fix circular import in artifacts
