+++ src/lib/server/job-executors/release.ts
@@ -1,8 +1,8 @@
 import type { Job } from 'bullmq';
 import { readFile } from 'fs/promises';
 import { CodeBuild } from '../aws/codebuild';
-import { AWSVars } from '../aws/vars';
 import { S3 } from '../aws/s3';
+import { AWSVars } from '../aws/vars';
 import { BullMQ, getQueues } from '../bullmq';
 import { prisma } from '../prisma';
 import { Build } from '$lib/server/models/build';
